+++ runtime-kotlin/runtime-application/src/test/kotlin/skillbill/application/FeatureTaskRuntimeRunnerTest.kt
@@ -5320,6 +5320,8 @@ internal class RuntimeFakeDatabaseSessionFactory(
 
   override fun <T> read(dbOverride: String?, block: (UnitOfWork) -> T): T = block(unitOfWork())
 
+  override fun <T> selfManagedWrite(dbOverride: String?, block: (UnitOfWork) -> T): T = block(unitOfWork())
+
   override fun <T> transaction(dbOverride: String?, block: (UnitOfWork) -> T): T {
     transactionDbOverrides += dbOverride
     return block(unitOfWork())